Conditions for new high or new low and Gann gap
Initiate a trade if and only if a new high or new low is reached during the OK to trade hours.  New highs or new lows before 8:00 am Eastern don't count.  Require that the gap between the new high (or new low) is a minimum of seven ticks from the entry price, whether entry is at the 25%, 50%, or 75% Gann level.